+++ Sources/HostBox.Borderline/IAssemblyLoader.cs
@@ -1,0 +1,11 @@
+﻿using System.Reflection;
+
+namespace HostBox.Borderline
+{
+    public interface IAssemblyLoader
+    {
+        Assembly Load(AssemblyName name);
+
+        Assembly LoadDefault();
+    }
+}