+++ RELEASE_NOTES.md
@@ -1,4 +1,13 @@
-﻿## 2.0.0-core1
+﻿## 2.0.0-core6
+  Extends IConfiguration interface by GetSection method.
+
+## 2.0.0-core4
+  Added a method to IAssemblyLoader returning default assembly.
+
+## 2.0.0-core2
+  Making IComponentFactory able to accept component configuration.
+
+## 2.0.0-core1
  - Migration to .net standard 2.0.
 
 ## 1.0.0